+++ packages/react/src/greyVest/index.js
@@ -9,19 +9,19 @@ export Fonts from './Fonts'
 export Icon from './Icon'
 export IconButton from './IconButton'
 export LinkButton from './LinkButton'
-export DropdownItem from './DropdownItem' // formerly also ListGroupItem
+export DropdownItem from './DropdownItem'
 export Modal from './Modal'
 export PagerItem from './PagerItem'
 export RadioList from './RadioList'
 export let SearchTree = () => {}
 export Select from './Select'
-export Style from './Style' // formerly GVStyle
+export Style from './Style'
 export * from './Tabs'
 export Table from './Table'
 export TabList from './TabList'
 export Textarea from './Textarea'
-export TextHighlight from './TextHighlight' // formerly Highlight
-export TextInput from './TextInput' // formerly Input
+export TextHighlight from './TextHighlight'
+export TextInput from './TextInput'
 export ToggleFiltersButton from './ToggleFiltersButton'
 export ToggleFiltersHeader from './ToggleFiltersHeader'
 export TreePauseButton from './TreePauseButton'
@@ -42,19 +42,3 @@ export TagsInput from './TagsInput'
 export Tag from './Tag'
 export DateInput from './DateInput'
 export StripedLoader from './StripedLoader'
-
-// removed exports
-// ----------------
-// Adder (now generic FilterAdder)
-// CheckButton
-// Pager (now vanilla ResultPager exampleType)
-// PagedResultTable (now a new exampleType)
-// UnmappedNodeComponent (now theme only)
-// FilterList
-// AddableFilterList (now part of SearchFilters)
-// FiltersBox (now part of SearchFilters)
-// QueryBuilder
-// SearchLayout
-// SearchFilters (now a generic search interface)
-// FilterButtonList
-// QueryWizard